Move the peer identity and pairing tests into their own file
Who a peer is and when we talk to one: free-agent pairing, BUSY and its
cooldown, staleness, and the five realm-canonicalization suites that keep
two same-name characters on connected realms apart.
